existing unit tests updated and passed.
