+++ mentorship ios/Views/Registration/SignUp.swift
@@ -10,6 +10,22 @@ struct SignUp: View {
     var signUpService: SignUpService = SignUpAPI()
     @ObservedObject var signUpViewModel = SignUpViewModel()
     @Binding var isPresented: Bool
+    
+    // use service to sign up
+    func signUp() {
+        // set inAcitivty to true
+        self.signUpViewModel.inActivity = true
+        
+        // make sign up api call
+        self.signUpService.signUp(
+            availabilityPickerSelection: self.signUpViewModel.availabilityPickerSelection,
+            signUpData: self.signUpViewModel.signUpData,
+            confirmPassword: self.signUpViewModel.confirmPassword
+        ) { response in
+            self.signUpViewModel.update(using: response)
+            self.signUpViewModel.inActivity = false
+        }
+    }
 
     var body: some View {
         NavigationView {
@@ -46,18 +62,7 @@ struct SignUp: View {
 
                     //sign up button
                     Button("Sign Up") {
-                        // set inAcitivty to true
-                        self.signUpViewModel.inActivity = true
-                        
-                        // make sign up api call
-                        self.signUpService.signUp(
-                            availabilityPickerSelection: self.signUpViewModel.availabilityPickerSelection,
-                            signUpData: self.signUpViewModel.signUpData,
-                            confirmPassword: self.signUpViewModel.confirmPassword
-                        ) { response in
-                            self.signUpViewModel.update(using: response)
-                            self.signUpViewModel.inActivity = false
-                        }
+                        self.signUp()
                     }
                     .buttonStyle(BigBoldButtonStyle(disabled: signUpViewModel.signupDisabled))
                     .disabled(signUpViewModel.signupDisabled)
