+++ mentorship iosTests/MentorshipTests.swift
@@ -8,37 +8,6 @@ import XCTest
 import Combine
 @testable import mentorship_ios
 
-class MockURLProtocol: URLProtocol {
-    static var requestHandler: ((URLRequest) throws -> (HTTPURLResponse, Data))?
-    
-    override class func canInit(with request: URLRequest) -> Bool {
-        return true
-    }
-    
-    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
-        return request
-    }
-    
-    override func startLoading() {
-        guard let handler = MockURLProtocol.requestHandler else {
-            XCTFail("Received unexpected request with no handler set")
-            return
-        }
-        do {
-            let (response, data) = try handler(request)
-            client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
-            client?.urlProtocol(self, didLoad: data)
-            client?.urlProtocolDidFinishLoading(self)
-        } catch {
-            client?.urlProtocol(self, didFailWithError: error)
-        }
-    }
-    
-    override func stopLoading() {
-    }
-}
-
-
 class MentorshipTests: XCTestCase {
     var urlSession: URLSession!
     
@@ -55,7 +24,7 @@ class MentorshipTests: XCTestCase {
     
     func testRequestAction() throws {
         // Set mock json for server response
-        let mockJSON = ResponseMessage(message: "response message")
+        let mockJSON = RequestActionResponse(message: "response message")
         // Create mock data from mock json
         let mockData = try JSONEncoder().encode(mockJSON)
